+++ dev/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.108](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.107...dev-web@0.8.108) (2024-08-18)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.8.107](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.106...dev-web@0.8.107) (2024-08-17)
+
+
+### Features
+
+* ✨ field 支持隐藏域 ([a0e8c15](https://gitee.com/newgateway/vtj/commits/a0e8c1586274004c19d052bf3e423d08887ad122))
+
+
+
+
+
+## [0.8.106](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.105...dev-web@0.8.106) (2024-08-17)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.8.105](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.104...dev-web@0.8.105) (2024-08-16)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+## [0.8.104](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.103...dev-web@0.8.104) (2024-08-16)
+
+**Note:** Version bump only for package dev-web
+
+
+
+
+
+
+## [0.8.103](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.102...dev-web@0.8.103) (2024-08-14)
+
+
+### Features
+
+* ✨ picker 支持格式化数据 ([882848d](https://gitee.com/newgateway/vtj/commits/882848d4a2936f669261dd0e64f8d13c34544e50))
+
+
+
+
+
 ## [0.8.102](https://gitee.com/newgateway/vtj/compare/dev-web@0.8.101...dev-web@0.8.102) (2024-08-13)